backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.1
numpy==1.23.5
opencv-python==4.7.0.72
pillow==9.3.0
torch==2.0.1
transformers==4.30.2
google-generativeai==0.3.1
face-recognition==1.3.0
psutil==5.9.0
ultralytics==8.0.145
soundfile==0.12.1
gtts==2.3.2
